refactor(components): refactor check-tag
