+++ packages/components/check-tag/index.ts
@@ -1,5 +1,7 @@
 import { withInstall } from '@element-plus/utils/with-install'
-import CheckTag from './src/index.vue'
+import CheckTag from './src/check-tag.vue'
 
 export const ElCheckTag = withInstall(CheckTag)
 export default ElCheckTag
+
+export * from './src/check-tag'
